plus::string:  Factor the data copy out of each code path in assign().
Whether we're writing to a freshly allocated large-string buffer or the built-in
small-string one, we have to (a) copy the data and (b) zero the terminator byte.
Set new_pointer to the buffer address in either case, and then operate on that.

Also, copy first, *then* zero the terminator byte, in case we are assigning part
of our own small-string buffer to itself.