Genie:  Remove superfluous FSTree_Virtual_Link.hh file.
I hesitate to use the word 'obsolete', as that would mean the file is no longer
relevant, implying that it once *had* been, which is not the case.

This file was untracked until the mass subfolder rename from FileSystem to FS,
at which point it was unintentionally committed with all the others.