HTTP:  Fix check for first "\r\n" in HTTP message
On no match, string.find() returns npos ((size_t) -1) -- which is a true
value.  That would leave itsStartOfHeaderFields set to offset 1 into the
message header.  Nothing good can come of this.

The only reason this hasn't shown up is that in practice, the first line
is sent in one packet -- but this isn't guaranteed to be the case.  That
can be demonstrated:

    http-server.vx:

    |   let CRLF = "\r\n"
    |
    |   let status = "HTTP/1.0 200 OK"          CRLF
    |   let ctype  = "Content-type: text/plain" CRLF
    |   let body   = "Hello world\n"
    |
    |   for c in status ctype body do
    |   {
    |       OUT <== c
    |
    |       sleep 0.1
    |   }

    $ listen :8080 minivx http-server.vx

    $ htget http://localhost:8080

This will work with this patch applied, and fail without it.

(Also, we replace string.find() with strstr() -- the original purpose of
this patch.)