poseven:  Rewrite no-arg gethostname() to return a pointer to static storage.
The client may but is no longer required to construct a std::string to store the
function result.

It's not reentrant, but the only way this could be an issue is for one thread to
call gethostname() and partially access the result buffer, then for the hostname
to be changed, then for gethostname() to called again by another thread (running
in the same process), and finally for the first thread to further access part of
the buffer.