plus::var_string:  Add private embiggen() member for expanding the string.
The embiggen() method is the one-stop shop for all your string expansion needs.
Appends?  Done.  Insertions?  No problem.  All in super-CPU-efficient amortized
constant time.  Not powerful enough for you?  Specify a capacity of your choice,
and embiggen() will reallocate the string to match -- even shrinking it, if you
tell it to.  Though if you request a capacity that's smaller than the existing
size, embiggen() uses the size instead, protecting you from your own mistakes so
you're free to code without having to think about such worries.

Here's how it works:  Pass in a new size, and if it exceeds the current capacity
then embiggen() will pick a new capacity for you -- either equal to the new size
or double the old one, whichever is greater.  Otherwise it does nothing.
Or, pass in a size and capacity, and embiggen() will use that capacity exactly
(except pinned to the size as a minimum), even if it shrinks the buffer -- ideal
for reserve().

Still not convinced?  The embiggen() method reallocates only if it needs to, and
then it copies your existing data to the new buffer -- completely automatically.
It even sets the new length.

Call now for a quote.[1]  If you're *still* not sold, try our 30-day, throwback-
guaranteed, double-or-nothing-or-your-money-back trial offer (specially tailored
to prise the goods from your tight fist -- at least long enough for the check to
clear).  No strings attached, of course.

Embiggen:  It's perfectly cromulent.[TM]

[1] Here's one:  "There's a sucker born every minute."  -- P. T. Barnum