Genie:  Support mmap() of /dev/zero.
Actually, this works on any simple device (/dev/{console,null,zero}) -- since
this is the Simplest Thing That Could Possibly Work.  Of course, it's only
actually *supported* for /dev/zero.

A possible future direction is supporting /dev/null for mapping memory that
isn't (necessarly) zeroed, but only if performance warrants it.