Genie:  Use temporary memory for large mmap requests through /dev/zero.
Try the application heap for anything under 128K, and otherwise (or if it fails)
use temporary memory.  In the latter case, explicitly zero the returned memory.

The cutoff is not 64K because dlmalloc makes requests of that size, and we don't
want *every* program spilling into the Process Manager heap -- especially jgetty
and sh.