nexus.pl:  A script to copy an MPW subset to a RAM disk.  Initial check-in.
The nexus.pl script looks for a RAM disk, renames it to 'Nexus', and copies over
files from MPW, including ToolServer and a subset of Interfaces and Libraries.
Also copied is part of the CodeWarrior for MPW distribution, which is expected
to be present.

We skip headers and libraries which are specific to MPW-provided compilers, MPW
tools, and even parts of CodeWarrior that we don't use, like SIOUX and two-byte-
integer libc.  Also omitted are deprecated headers from Universal Interfaces,
like Types.h.