Genie:  Call OTListen() with the endpoint in nonblocking mode.
If the socket is blocking, yield until the endpoint is in other than idle state.
This will normally be T_INCON (incoming connection pending) but could perhaps be
T_UNINIT.  In either that case, or the idle state with a nonblocking socket, the
subsequent call to OTListen() will throw an exception.

In the event of an exception, the endpoint is left in nonblocking mode, but the
only useful things to do with the socket at that point are another accept() call
(which is only going to set it nonblocking anyway) or close(), which not only is
unaffected, but moots the issue permanently.

One major benefit of this change is that accept() now idles without entering the
notifier function (through which we can't throw exceptions), so now we can throw
EINTR when a signal handler completes to interrupt accept().