A-line:  Limit mxcpp #line directives to MacRelix
Currently, mxcpp relies on passing REALPATH_OUTPUT_HFS to _realpathat()
in order to get the Mac pathnames that MWC can process without crashing.
Providing an alternate implementation using Carbon routines is possible,
but (a) annoying (for creating a dependency on Mac OS) and (b) won't work
on Linux.  Instead, if we're building for MacRelix on a non-MacRelix
platform, just punt the issue and omit #line directives entirely.  Any
subsequent compile diagnostics will refer to the preprocessor output, but
this can be remedied by rerunning the preprocessor locally.