plus::string:  Add protected copy_on_write() member.
The copy_on_write() method either assigns a fresh copy of the string's contents,
or does nothing if the string is already suitable for modification, according to
these rules:

Small strings are self-contained and therefore ready for modification.

If the actual capacity isn't known, assign a new copy.  This can only occur with
handoff assignments, where specifying a capacity is optional, and bears semantic
significance:  If specified, then the string may be modified, and otherwise not.
So this is not merely an issue of not knowing the capacity -- a capacity of zero
*means* that the string may not be modified.

If this isn't a shared string, then we own the buffer and can modify it at will.

Similarly, we own the buffer if the reference count is one.  Otherwise, reassign
the string's contents.