[FIX] plus::string:  Fix copy-on-write memory corruption
The call to assign() in copy_on_write() was eliminating excess capacity.  As a
result, appending data small enough to fit the capacity (as measured before the
call to copy_on_write()) would overrun the reallocated buffer.

Add an optional capacity parameter to pointer/length assign(), and pass in the
current capacity from copy_on_write().