Genie:  Make symlinks, and also store the target in aliases
Make fake a symlink (a file of type 'slnk' and creator 'rhap', containing
a target pathname, like OS X symlinks but with a regular file mode) when an
alias can't be made (because the target doesn't exist or isn't in the File
Manager's domain).

Set the Shared bit on both aliases and (fake) symlinks to indicate that the
path is stored as MacRoman, in case we switch to UTF-8 in the future.