Genie:  Check __MACOS__ when guarding against Carbon (for OS X)
We don't define TARGET_API_MAC_CARBON on OS X anymore (since Carbon is always in effect) which means
that it isn't defined at all until TargetConditionals.h is included.

Check __MACOS__ first.  Only if it's defined are we targeting classic Mac OS, in which case we check
TARGET_API_MAC_CARBON (which will be defined to 1 for Carbon builds).