Nitrogen:  Remove CFData.hh dependency on <vector>
Rewrite functions accepting and returning vector< UInt8 > values as templates.  Calls to the latter
must be rewritten to specify the template parameter.  Calls to the former should work as is.

There are two semantic differences as well:  One, we don't substitute NULL when the length is zero.
There's no reason to do this, and it's already been observed to be harmful in (I think) the case of
FSRead(), which returns paramErr given a NULL buffer.  Two, the deliberate conversion to CFIndex is
elided, though this invocation actually hasn't done anything since the specializations were moved to
convert_numeric.hh (which isn't included) anyway.

I haven't decided what should be done (if anything) to check for negative length values.  First I'll
need to see if the OS already has such checks.  If an explicit check is to be reinstated, it will be
a simple one, not a generic catch-all -- convert() is for templates.