[FIX] MacFixup:  Add Carbonate prerequisite (and remove its unneeded MacOS to break a cycle)
MacFixup now depends on Carbonate.

MacOS is concerned with linking against the right libraries and searching FlatCarbon on OS X.  The
former is not an issue as long as MacOS is used by some project (which it is, since MacOS is what
uses Carbonate (through MacFixup)) and the latter is moot since Carbonate isn't used in OS X builds.