tools:  Don't link microtools against MacOS or MSL-fixup
The goal is to avoid linking against libraries that differ between Carbon
and non-Carbon when they're not needed.  The linker recognizes when neither
InterfaceLib or CarbonLib are supplied and indicates the binary as being
Carbon-agnostic (though this is arguably an error, as MathLib is morally
equivalent to InterfaceLib).

Microtools don't use MSL's malloc (which is enforced my nomalloc hiding the
symbols), so they don't need to clean up the malloc pool (which requires
DisposePtr()).  They also don't use stdio, so the rest of MSL-fixup doesn't
apply either.

TimeOff, used by localtime[_r](), calls ReadLocation().

Kerosene's clock() calls Microseconds(), so any microtool using it (none,
so far) must declare MacOS as a prerequisite.

We need to use MacFixup so our fixup headers are found instead of Apple's.