recall:  Avoid dynamic allocation while crawling the stack
Remove vector from recall's interface.  Pass array and length instead of
a vector reference to make_stack_crawl(), and return the frame count.

(In addition to smaller translation units, this lets the caller choose a
maximum frame count (which remains 64 in all cases, for now).)

Store stack frames in a fixed-length array in debugging_context, which is
(a) guaranteed not to throw (required for objects thrown as exceptions),
and (b) more efficient anyway.