Genie:  Store a process group ID, not weak pointer, in Terminal
Store a terminal's foreground process group as an ID number instead of a
weak pointer.  The null value is 0x7fffffff, suitable for tcgetpgrp().

In the ProcessGroup destructor, check if the session's controlling terminal
(if any) has 'this' as its foreground process group, and if so reset its
foreground process group to the null value.

Check for null proc.GetProcessGroup() in SendSignalToProcessGroup().