Nitrogen:  Remove FSForkRef and company
There's no real need for it.  If it turns out to be necessary (e.g. for
managing temp files as owned< FSForkRef >) then it should be a template
with the fork name type as a parameter -- there's no sense dragging in
basic_string when the fork name is likely to be a constant or even null.