Genie:  Convert UTF-8 symlink targets to MacRoman (Carbon)
Mac OS X symlinks store the target path in UTF-8, which must be converted
to MacRoman.  Fake symlinks and aliases with a data-fork-stored target path
will have the Shared bit set to indicate MacRoman encoding; its absence
indicates UTF-8 (as is the case with real symlinks).

For now, this is Carbon-only.  Non-Carbon Lamp is not compatible with UTF-8
symlink targets containing non-ASCII characters.