compat/lamp:  Add restack unit for _lamp_restack()
Define struct _lamp_stack_footer (for both architectures) and declare
_lamp_restack().  Define a dummy _lamp_restack() for non-Lamp since gcc
fails to elide code that's trivially observed to be unused when optimizing
is off.

A Lamp user stack is a block of memory whose end is described by
struct _lamp_stack_footer.  This structure includes a properly-aligned
stack frame (large enough to call an 8-argument function, in the case of
PowerPC) as well as a pointer to the beginning of the block (which is also
the stack limit) and a pointer to the previous user stack (or NULL).