[FIX] Genie:  Resolve $port/window correctly
This was probably broken by ceef596[1], which no longer allows the working
directory to be a symlink.

An alternative fix is to insert { ref = ref->ResolveLink(); }.  This saves
resolving the link redundantly but at the cost of adding additional code.
The difference is likely to be microseconds, not milliseconds, so we prefer
data (in the pathname) to code.

[1] ceef596 Genie:  Call ResolvePath() once before the loop