libc:  Use TRAP #0 for system calls
Remove the central user-side dispatcher and replace branches to it with
trap instructions.

The special case for system call numbers between 80 and FF is now obsolete.
Previously we factored out the ANDI instruction, allowing us to use a MOVEQ
instead of the larger MOVE in those stubs.  But now that difference (of two
bytes) is outweighed by the difference between six bytes of code to jump to
the user-side dispatcher vs. just two to trap directly.  In other words, it
now costs more to have a central user-side dispatcher at all.