plus:  Make struct freeze_pod a template
The issue is single-byte fields.  The ones in Genie's /sys/mac/unit already
derive from serialize< char > but override result_type to be int, which is
mildly more convenient in a debugger vs. char which would require a manual
step to display as integer.

Member function templates were being instantiated for int instead of char.