MacBinary:  Clear kHasBeenInited on the root item
Clear the inited flag for just the top-level item, regardless of whether
it's a file or directory.

This is not a change for single-file archives, since the item is always a
file (and the flag was cleared) and it's always the top item (and the flag
will be cleared).  But for folder archives, this clears the inited flag on
the top-level folder (which before it didn't) and doesn't clear the flag on
files (which before it did).