privileged:  Add 'machine 68020' to compile MOVEC
This should cause an illegal instruction exception on a 68000 Mac, which if
properly handled will result in raising SIGILL -- which somewhat ironically
is identical to the behavior in the correct case (at least until siginfo is
implemented).

The problem with things like touching SR is that the user-mode environments
in Mac OS (i.e. 68K boot with virtual memory, 68K emulated) don't propagate
the exception into user code, but instead behave as if no violation occurs.
Furthermore, the lack of a system-supported user-mode context in Mac OS for
the 68000 implies that Lamp will always be running in supervisor mode, so a
test for privilege violations would do nothing -- unless Lamp itself turned
off supervisor mode, in which case the implementation of such feature would
be a prerequisite for testing it.  In the meantime, this program will crash
a 68000 Mac, requiring a reboot (at least in System 7.1, with no debugger).