Genie:  Rethrow Gestalt() errors in ioctl()
Rethrow gestaltUndefSelectorErr (the culprit) as EINVAL, but also check for
gestaltUnknownErr (and rethrow as ENXIO).

POSIX says that ioctl() may yield EINVAL if the request (in /dev/gestalt's
case, the selector) is invalid, and ENXIO if the argument is valid but "the
service requested cannot be performed"[1], which seems to be an acceptable
conclusion "if Gestalt doesn't know the answer"[2].

[1] http://www.opengroup.org/onlinepubs/009695399/functions/ioctl.html
[2] MacErrors.h

Also, call stock Gestalt() instead of using Nitrogen.