Genie:  Add support for 4-char codes in filenames
Strip trailing spaces, or hexify the whole thing if any bytes are unsafe.
Safe characters include alphanumerics and underscore, and as a special case
the last char may be '#' (as in 'ICN#') since it's prevalent.

(Some other punctuation characters that have been seen are '!' and '&',
which are out of the question due to being shell-unsafe.)