[FIX] A-line:  -d aeut_RezTemplateVersion on OS X
Fix regression of aeut_RezTemplateVersion to default of 0 using OS X's Rez.

For whatever reason, we're using the 'wide version', version 1.  We define
aeut_RezTemplateVersion to 1 and include AEUserTermTypes.r in Genie.aete.r,
but A-line's include of Types.r (the FlatCarbon version of which includes
Carbon.r and thence AEUserTermTypes.r) defeats this.  So explicitly define
aeut_RezTemplateVersion to 1 for Mach-O builds before including Types.r.

These comments apply:

	What is aeut_RezTemplateVersion for?  I can't find any documentation
	about this.

	Why are there two different formats?  What are they for?  When would/
	should I use one vs. the other?  When was version 1 introduced and why?

This was the closest attempt to an answer:

	So, presuming that a wide array is unpacked and packed otherwise, I
	think that would explain it.  I'm not entirely sure what "wide array"
	means, and finding an online rez language reference manual seems to be
	impossible.

http://lists.apple.com/archives/applescript-implementors/2002/Nov/msg00008.html
http://lists.apple.com/archives/applescript-implementors/2002/Nov/msg00016.html
http://lists.apple.com/archives/applescript-implementors/2002/Nov/msg00015.html