Genie:  Remove register-saving code for ptrace()
We aren't using it, and it would need to get rewritten anyway.  The PowerPC
registers can be moved with one instruction, the scratch registers aren't
important, and the actual 68K registers at the time of the system call can
be saved in the TRAP exception handler if desired.  PPC could be similarly
treated if we used the sc instruction, but that wouldn't work in Rosetta or
SheepShaver, though otherwise the registers are useless, unless perhaps we
use assembler code called from the dispatcher to save them.