Genie:  Release the old executable in same thread
Release the reference to the current executable in Exec() before returning
to execve() instead of waiting until the new thread runs.  For the vfork()
case, the caller has the same executable, though it might now be disposed
earlier if the caller exits immediately after execve() returns, which is a
bit unusual though certainly possible.  Otherwise, the thread is terminated
in execve() and no return to the userland code is possible.