libc:  Remove crufty alloca() declaration
First of all, if <alloca.h> is included before <stdlib.h>, then alloca() is
defined as a macro and __alloca() is declared.  The second problem is that
alloca() is conflictingly declared as taking an int parameter rather than
size_t.

Just remove the mess and leave the <alloca.h> include.