ld:  Don't expedite Orion
Before we had to make sure that main() was linked early so it would be
reachable by a 16-bit jump.  This was no issue for non-Orion-clients, which
defined main() in a unit typically linked first, but Orion defines main()
in a library (which is linked later).  Anyway, _lamp_main() is now set as
the main entry point directly, so main() is no longer relevant.