jsync:  Don't pass owned< fd_t > when fd_t will do
Passing owned< fd_t > objects as arguments requires the receiver to include
code for destructing them but doesn't absolve the caller of the same duty.

Behavior-wise, the difference is that the object gets destroyed immediately
after the called function returns instead of immediately before the called
function returns.  In other words, it's an implementation detail.