Genie:  Add unit for resources as files in a dir
This is different from the existing implementation in several ways:

* The resources are all mapped into single directory, instead of subdivided
by resource type
* The filenames end with the type following a dot, so extension-based file
associations will work
* The ID is expressed in hexidecimal for easy sorting[1]

[1] Sure, you could pad a decimal field to five digits, but you still have
to treat IDs as unsigned, because a filename starting with '-' is asking
for trouble, and so is a decimal number starting with 0, as is a filename
beginning with a space.  Hexidecimal for the win.