OSErrno:  Don't map eofErr to zero
FSpOpenResFile() returns eofErr when the file exists but the resource fork
is empty.  At least one caller wants ENOENT in this case, but we're not
about to establish such a mapping for all cases.

The eofErr returned from FSRead() and such already have to be handled
specially (for POSIX short read semantics) and thus aren't affected.