Genie:  Add an empty resource if needed on O_CREAT
Immediately add an empty resource on open() with O_CREAT if one doesn't
already exist.  Otherwise, it will stat as nonexistent until the resource
is written.

This fixes an issue with perl, where fstat() on a created resource yielded
ENOENT, resulting in failure to create a new resource via open.