Genie:  Return to system call dispatcher (68K)
Don't just jump to the system call, but since we're not duplicating the
arguments, it's more complicated than just changing JMP to JSR.

We can't push our return address without messing up the stack, so we have
to replace our caller's return address -- which means we have to save it
somewhere, and we can't just push it onto the stack or we would have done
that with our own return address in the first place.

Save the caller's return address into A4 and overwrite the stack top with
ours (the address of the instruction following the JMP).  On return, push
the caller's return address onto the stack (since our own was popped when
the system call performed its RTS).  Restore A4 from user_pb.globals.