[FIX] Genie:  Don't clobber signal handler in A4
When calling a signal handler, don't set userspace A4 until we know that it
doesn't hold the signal handler address.  This fixes a crash when calling
signal handlers in 68K, since the handler address *is* stored in A4 and
gets clobbered with the globals address.

Replace the macro-bracketed call-through-pointer with a call to a function
taking the handler address and signo in registers A0 and D0 (on 68K).  Then
we're free to clobber A4 temporarily since the handler is in A0.