default_value_traits::value() -> null_resource::get()
The function null_resource< Resource >::get() returns a value suitable for
initializing an ownable resource to its null value, i.e. the value for that
type indicating that it's not live and there's no resource to release.

The unspecialized definition returns a default-constructed object, suitable
for fundamental types and enums.  It's also usable for classes defining the
default constructor, although this will result in some code bloat with poor
optimizers.  In any case, aggregates with no constructors specified require
a specialization of null_resource.