Genie:  Rewrite DeliverFatalSignal() to be robust
Rewrite DeliverFatalSignal() to retrieve the signal action and either call
the signal handler directly or terminate the process immediately -- or both
if the handler returns.

This doesn't protect against recursively fatal signal handlers (such as a
SIGILL handler that executes an illegal instruction) but it avoids letting
some other signal take precedence.