Genie:  Call Suspend() in Terminate()
Instead of manually assigning NULL to global_parameter_block.current_user
and gCurrentProcess (and doing so only for non-vforked processes), call
Suspend(), which does that and also updates the system timer.  However,
proceed only if this is the current process.

Remove the Forked() test and call itsThread.reset(), which does nothing for
vforked processes since by definition Forked() is true only if itsThread is
null.  For non-vforked processes, reset() will not return (since it results
in the thread's termination) but it clears itsThread before that happens.