Genie:  Fix OTSocket::SysWrite() to return total
Return the total number of bytes written, rather than the number written by
the last call to OTSnd().

No known bugs have occurred as a result of this error to date.  In the case
of non-blocking sockets, only one call is ever made to OTSnd() in the first
place.  Blocking writes may have returned a short value, but it's customary
to ignore the return value (beyond checking for failure) unless signals are
caught, since in that case writes are guaranteed to succeed fully.  Even in
the event the return value is checked, it's only an issue when a short send
occurs, which in my brief test of Open Transport 2.7.9 it didn't, even when
I sent 65536 bytes in a single call.