Genie:  Don't read further if peeked data present
If the peek buffer contains data from a prior Peek() call, don't also call
SysRead().  Catching exceptions will be inadequate once a caught signal can
be thrown (since the signal will be lost if peeked data exist).

In practice this has no effect:  The only user is pump(), which peeks at a
minimum of one byte, writes the entire peek buffer, and advances past it.
In the event of a partial write, only an equal part of the peek buffer gets
skipped.  The subsequent peek will return the same buffer, and Read() will
return the balance remaining instead of filling an entire 4K block with the
remainder plus the result of a fresh read.  In the case of reading from the
disk and writing to the network, this theoretically may improve performance
in the event of a caught signal, although a corroborating benchmark is not
imminent.