Pedestal:  Rewrite WakeUp() and move to its own unit
Set gInWaitNextEvent to true while we're calling WaitNextEvent().  After we
set it but before calling WaitNextEvent(), check gAsyncEventsReady; if true
(which will be the case if an asynchronous operation completes while we are
the current process) then set it to false and set sleep to zero.

In WakeUp(), check gInWaitNextEvent; if false, then our process is current,
so set gAsyncEventsReady to true; otherwise, call WakeUpProcess(), which if
successful means we're done and if not means we have to try later, since it
may be too late to set gAsyncEventsReady, and it's too early to wake up our
process since it's still current.