Annex:  Define {P,S,U}RefCon for pre-10.5
Define PRefCon, SRefCon, and URefCon for compatibility with 64-bit Carbon
(if they're not already defined).

(Notice that "compatibility" refers to "64-bit Carbon", not "64-bit" in
general.  The types void* and long are already 64-bit compatible, and no
type aliases are required, but in 64-bit Carbon, SRefCon switches from long
to void*, and a name is needed to refer to it in all cases.