Lamp:  Rename jgetty to teletype
The name jgetty is problematic with respect to both the 'getty' root and
the 'j' prefix.  Rather than a getty program that runs for a tty while it's
not in use and then passes the torch to login, jgetty acts more like xterm,
in that it runs only when a tty is in use.  The 'j' initial is short for
'jTools', which we haven't ruled out generally but won't be using for
MacRelix.

Given MacRelix' distinction between a terminal and a tty, use of the former
is imprecise, since a terminal refers to any interface boundary between the
user and a set of processes (i.e. a session, in POSIX), whereas a tty is a
specific form of terminal -- one whose interface resembles a teletype, in
that it sends text that you type and displays text that it receives.  Since
the name 'tty' is taken, and it's more user-friendly to spell things out
than use cryptic abbreviations, and also because the anachronism of it just
appeals to me, the new name is 'teletype'.