sh:  Clear out stale PWD and OLDPWD starting shell
Unset the PWD environment variable before the shell executes "cd ." to set
it.  Otherwise the logical path bookkeeping gets confused and leaves PWD as
"/" if it's already set to that, even if the actual working directory is
something else.

Unset OLDPWD as well, since its value (if set) is necessarily stale, and
export the unset variable as bash does.

This fixes a bug where PWD remains "/" though the working directory matches
HOME when login is run from a shell script.