tlsrvr:  Quote {MPW} for when it contains spaces
Delimit {MPW} with double quotes when passing it to ToolServer so it won't
get tripped up by pathname components with spaces (or other metacharacters,
for that matter).

This is mostly an issue with volume names, such as "Macintosh HD", which
Lamp typically avoids (since the pathname to a volume contains its volume
reference number (inverted) rather than the volume name), but to which
ToolServer is susceptible.