Genie:  Fix integer overflow in nanosleep()
A sleep time of longer than 0xFFFFFFFF/60 seconds got truncated when passed
to Ped::AdjustSleepForTimer().  The sleep argument to WaitNextEvent() is
signed, so sleep can't exceed 0x7FFFFFFF ticks.  If our sleep parameter is
longer, then there's nothing to do, since sleep is already limited beyond
what we require.

This particularly affected doze(), which defaults to a maximum of a billion
seconds.