MacRelix:  Don't overconvert /sys/mac/user/name
We were mistakenly using sys_mac_user_name as if it were MacRoman-encoded,
which it actually (mistakenly) was -- in classic Mac OS, that is.

Map the .mac-name file to sys_mac_user_macname.  Add the missing conversion
to sys_mac_user_name and remove the explicit conversion in the mapping of
name.