plus:  Ignore shared capacity in copy_on_write()
That is, check for a delete_shared policy before a non-zero capacity.

A zero capacity is used for externally created buffers that we can't touch,
and therefore (currently) implies that the buffer isn't shared.  But in the
future we might want to use the 'capacity' of a shared buffer in some other
way, such as for substrings.