plus::string:  Initialize to empty string in ctors
Instead of just setting the last cell to 0, which establishes a valid but
uninitialized string of 15 characters, properly initialize the buffer to an
empty string, except in the case of the fill constructor, where there's no
benefit.

The pointer/length assign() member works directly on empty strings, but
otherwise calls itself recursively on a temporary and swaps it with *this.