plus::string:  Move 'privileged' reset() to string
Replace limited string::reset() with the more powerful var_string::reset().

The latter is simply a wrapper of reallocate(), accepting a length argument
and returning the buffer pointer.

This allows for creating strings with the pattern of allocating in full and
then filling in the data without using a var_string, which would be tainted
and require an explicit move() call to avoid the unnecessary buffer copy.