plus::string:  Remove delete_none
The idea was to build a string's contents in a stack buffer and then create
a string that uses that buffer, avoiding the need to reallocate the storage
while building the string, but also deferring dynamic allocation completely
until the string is copied (if ever).

Now we use reset() to allocate the string's buffer.  This requires a memory
allocation upfront, versus none if the string is only used locally, but has
the advantage of being movable and copyable shallowly, whereas local-buffer
strings can't even be moved safely without reallocation -- To move from one
local string to another is safe, but returning one via move() is not.

Now that the convenience issue is addressed by reset(), delete_none remains
as only a belatedly premature optimization.  Remove it.