Genie:  Track exec images by file ID, not FSSpec
The file ID hash is a combination of the file ID and the vRefNum:

	hash = vRefNum << 24 ^ fileID

The low byte of the volume reference number is XORed into the high byte of
the file ID, yielding the hash.

Since file/dir IDs are monotonically increasing, a collision wouldn't be
possible until after the 16 millionth file on a volume.  In the unlikely
event of such a collision, the two images would push each other out of the
cache, but still run correctly.  Note that this is only an issue if running
binaries from multiple volumes.