pfiles:  Use back() vs. end() and combine appends
Explicitly set { const bool has_trailing_slash = string.back() == '/'; }.
Use back() instead of the complex end() expression and save the result in a
self-documenting constant.

Use has_trailing_slash as an offset into the string to append, combining
the conditional slash with the next argument when it's a string literal.