osascript:  Don't pass n::string to p7::write()
The p7::write() templates are going away, trading the ability to pass a
nucleus::string (which we use only rarely, where we mix Mac OS with POSIX)
for the ability to pass an expression that converts to a plus::string (e.g.
a concatenation) which will be far more useful.

Also, we can tweak string-returning Nitrogen functions to accept a template
parameter for the result type.