recall:  Fix demangling edge case and add tests
Terminate a function parameter list at ',' as well as '>' (which happens if
the function is an extern template parameter and is followed by yet another
template parameter).

This unbreaks the following symbol:

.AEInstallEventHandler<&.Adapter__Q28Nitrogen110AEEventHandler_Callback<v,&.Handler__Q27UseEdit18GetData_AppleEventFRCQ23Mac11AEDesc_DataRQ23Mac11AEDesc_Data>FPC6AEDescP6AEDescl,PCv>__8NitrogenFQ23Mac12AEEventClassQ23Mac9AEEventIDPCvb

Add the above test and two others, plus additional code to parse tests with
input and output on separate lines (which are separated from other tests by
a double newline).