Nitrogen:  Fix async File Manager call error check
On 68K, ignore the function result of asynchronous File Manager calls, and
return the ioResult field instead.  For PBRead() and PBWrite(), which when
calling the File Manager require this fixup and calling the Device Manager
must be used normally, check the ioRefNum field.  It's negative for device
drivers and non-negative for files, so substitute the ioResult value only
in the latter case.

Non-68K doesn't have this issue and is unaffected.