plus::string:  Share buffers in substring assign()
Merge copy assign() and 3-arg substring assign() by making the pos argument
optional.

For small strings, take the path of reassignment versus shallow copy, since
the latter is now specific to external buffers.

For large strings, update the length member, and if the offset parameter is
non-zero, then overwrite the capacity member with the inverse of the offset
(stacked with any existing offset).

If reassignment takes place, add the pos offset to the input string's data.