plus::string:  Add substr_offset() and is_c_str()
Add trivial definitions of substr_offset() and is_c_str().  The former call
returns the number of bytes in the string buffer preceding the actual start
of the string value, which (until substrings share buffers) is always zero.
The latter indicates whether the string's data bytes are followed by a zero
byte (making a valid C string), which is always true at present.