plus::string:  Implement is_c_str() for real
Check for a NUL byte following the string data.  Note that how it got there
isn't important:  If a prefix substring omits from the original data a tail
that just happens to begin with a NUL, then the substring is a C string and
calling is_c_str() on it returns true.

Also note that it's possible to construct a string (using 4-arg constructor
or assign()) whose buffer fills an entire page of memory such that the next
byte is unaddressable.  In this case, the calling code must ensure that the
is_c_str() member is never called, or the results are undefined.  It's also
the caller's responsibility to guarantee that the byte following the string
won't become non-zero between a call to is_c_str() and the eventual release
of the string's buffer (either due to reallocation or destruction).  Said a
different way, once is_c_str() returns true it must continue to return true
for that string object (since reallocation will always yield a new string).