plus::string:  Implement substr_offset() for real
Overload the capacity member to store the substring offset (inverted).  The
actual capacity value is limited to the lower half of the address space, so
the high bit is never set (and signedness doesn't matter).  Reinterpret the
capacity member as signed, and interpret negative values as inverse offsets
(and zero (unspecified) capacity).