plus::string:  Fix _policy() when char is unsigned
Return signed char from the const _policy() accessor, so the result will be
evaluated as signed regardless of the signedness of char.

Make the non-const _policy() accessor take an argument instead of returning
a reference, since we must ensure that any value returned is signed.  While
we can't return a reference to signed char (without an evil and ugly cast),
we can remove the function entirely and replace it with an incompatible one
(requiring that we modify the call sites as well).

Ordinarily, it's considered better style to name a setter set_item() rather
than just item(), but we've deliberately preceded the word in question with
an underscore as a visual reminder that the delete_policy values have to be
negated (accomplished with a preceding '~'), and the name set_policy() just
doesn't have the same effect, since it's a typical construction.