plus::string:  Refine move() effect on source data
For large strings, set the last byte to max_offset (yielding a null string)
instead of zero (leaving a string of max_offset garbage bytes).

For small strings, don't munge the source at all, but leave it intact so as
to preserve the validity of pointers into the source string in all cases.